Revert "Dz4"
