--- lib/func/appin.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:homework3/UI/homepage.dart';
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
-      ),
-      home: const MyHomePage(title: 'Chukhlomina Ekaterina'),
-    );
-  }
-}
