+++ pubspec.yaml
@@ -59,10 +59,9 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  assets:
-     - assets/images/rfYH1DvF6u0.jpg
-     - assets/images/RgWXwPeKp1o.jpg
-     - assets/images/Фото 3х4 близко — копия.jpg
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
